Extended FVarIDs of SurfaceFactory from int to intptr_t:     - changed FVarID typedef to use intptr_t (signed use default of -1)     - added methods and member to assign default FVarID in Options     - overloaded face-varying Surface construction/initialization methods     - updated Factory subclasses for TopologyRefiner with explicit casts     - updated bfr/tutorial_1_2 to use dflt FVarID Option and new methods

+++ opensubdiv/bfr/refinerSurfaceFactory.cpp
@@ -60,7 +60,7 @@ RefinerSurfaceFactoryBase::~RefinerSurfaceFactoryBase() {
 inline int
 RefinerSurfaceFactoryBase::getFaceVaryingChannel(FVarID fvarID) const {
 
-    return ((0 <= fvarID) && (fvarID < _numFVarChannels)) ? fvarID : -1;
+    return ((0 <= fvarID) && (fvarID < _numFVarChannels)) ? (int)fvarID : -1;
 }
